Modified for possibility of 64 bit default integer size
git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@742 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3d49e66a48058f9097be5c8fd69c8b1801220707

+++ src/int2txt.f90
@@ -1,5 +1,5 @@
 !----------------------------------------------------------------------
-! $Id: int2txt.f90,v 1.3 2005-06-23 20:07:53 dschanen Exp $
+! $Id: int2txt.f90,v 1.4 2006-04-21 20:59:24 dschanen Exp $
 
 ! Program int2txt
 ! Take a binary file of 34 4 byte integers and generate text file from it
@@ -16,7 +16,7 @@ program int2txt
 implicit none
 
 integer i
-integer, dimension(34) :: x
+integer(kind=4), dimension(34) :: x
 
 open( unit=10, file='rand_seed.bin', action="read", recl=4, &
       form='unformatted', access='direct' )
